+++ packages/plugin-ext/src/common/plugin-protocol.ts
@@ -22,6 +22,7 @@ import { ExtPluginApi } from './plugin-ext-api-contribution';
 import { IJSONSchema, IJSONSchemaSnippet } from '@theia/core/lib/common/json-schema';
 import { RecursivePartial } from '@theia/core/lib/common/types';
 import { PreferenceSchema, PreferenceSchemaProperties } from '@theia/core/lib/common/preferences/preference-schema';
+import { ProblemMatcherContribution, ProblemPatternContribution, TaskDefinition } from '@theia/task/lib/common';
 
 export const hostedServicePath = '/services/hostedPlugin';
 
@@ -72,6 +73,9 @@ export interface PluginPackageContribution {
     keybindings?: PluginPackageKeybinding[];
     debuggers?: PluginPackageDebuggersContribution[];
     snippets: PluginPackageSnippetsContribution[];
+    taskDefinitions?: PluginTaskDefinitionContribution[];
+    problemMatchers?: PluginProblemMatcherContribution[];
+    problemPatterns?: PluginProblemPatternContribution[];
 }
 
 export interface PluginPackageViewContainer {
@@ -173,6 +177,27 @@ export interface PluginPackageLanguageContributionConfiguration {
     wordPattern?: string;
     indentationRules?: IndentationRules;
     folding?: FoldingRules;
+}
+
+export interface PluginTaskDefinitionContribution {
+    type: string;
+    required: string[];
+    properties: {
+        [name: string]: {
+            type: string;
+            description?: string;
+            // tslint:disable-next-line:no-any
+            [additionalProperty: string]: any;
+        }
+    }
+}
+
+export interface PluginProblemMatcherContribution extends ProblemMatcherContribution {
+    name: string;
+}
+
+export interface PluginProblemPatternContribution extends ProblemPatternContribution {
+    name: string;
 }
 
 export const PluginScanner = Symbol('PluginScanner');
@@ -365,6 +390,9 @@ export interface PluginContribution {
     keybindings?: Keybinding[];
     debuggers?: DebuggerContribution[];
     snippets?: SnippetContribution[];
+    taskDefinitions?: TaskDefinition[];
+    problemMatchers?: ProblemMatcherContribution[];
+    problemPatterns?: ProblemPatternContribution[];
 }
 
 export interface SnippetContribution {